+++ game-frontend/projects/game-admin/src/app/types/requirement-information-with-translation.type.ts
@@ -1,4 +1,4 @@
-import { RequirementInformation } from '@owge/universe';
+import { RequirementInformation } from '@owge/core';
 
 /**
  * Has frontend computed translations for the requirement