+++ game-frontend/projects/owge-universe/src/lib/storages/unit-type.store.ts
@@ -1,5 +1,5 @@
+import { UnitType } from '@owge/core';
 import { ReplaySubject, Subject } from 'rxjs';
-import { UnitType } from '../types/unit-type.type';
 
 
 /**